Update readme.md to include schema.sql
Probably because I am a Node newbie, but I ran into a trivial problem. Although there's a clear schema.sql file in the main directory, it took me a while to find it because I was just copy-pasting from readme.md to get my sessions going. Couldn't figure out the exact schema till I happened to see the file. Would be good if there are explicit instructions for it.